queue: split struct Queue into submodule
Split `struct Queue` into a submodule.
